Revert "Modular dev 1.0.2"
