+++ project/CMakeLists.txt
@@ -1,23 +1,9 @@
 cmake_minimum_required(VERSION 3.31)
-project(SLN2Code_Modular)
+project(SLN2Code)
 
 set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+project(SLN2Code LANGUAGES CXX)
 
-add_executable(SLN2Code_Modular src/utils.cpp
-        src/compiler_service.cpp
-        src/debugger_service.cpp
-        src/project_structure_service.cpp
-        src/sha256.cpp
-        src/library_service.cpp
-        src/project_generator.cpp
-        src/command_line_parser.cpp
-        src/builtin_library_provider.cpp
-        src/remote_library_provider.cpp
-        src/main.cpp)
-include_directories(include)
-add_executable(SLN2Code ${SOURCES})
-if(WIN32)
-    target_link_libraries(SLN2Code ws2_32)
-    add_definitions(-D_WIN32)
-endif()
-install(TARGETS SLN2Code DESTINATION bin)
+INCLUDE_DIRECTORIES(${PROJECT_SOURCE_DIR}/include)
+add_executable(src main.cpp)
