--- project/include/Logo.h
@@ -1,10 +1,0 @@
-#pragma once
-#include <iostream>
-
-inline void Logo() {
-    std::cout << "  ____  _     _   _ ____   ____          _      \n"
-              << " / ___|| |   | \\ | |___ \\ / ___|___   __| | ___ \n"
-              << " \\___ \\| |   |  \\| | __) | |   / _ \\ / _` |/ _ \\ \n"
-              << "  ___) | |___| |\\  |/ __/| |__| (_) | (_| |  __/ \n"
-              << " |____/|_____|_| \\_|_____|\\____\\___/ \\__,_|\\___| \n";
-}
