--- project/include/debugger_config.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "constants.h"
-#include <string>
-
-class DebuggerConfig {
-public:
-    std::string name;
-    std::string path;
-    DebuggerType type = DebuggerType::UNKNOWN;
-};
