--- project/include/directory_node.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <string>
-#include <vector>
-
-class DirectoryNode {
-public:
-    std::string name;
-    std::vector<DirectoryNode> children;
-};
