--- project/include/project_config.h
@@ -1,14 +1,0 @@
-#pragma once
-#include "compiler_config.h"
-#include "debugger_config.h"
-#include <vector>
-#include <string>
-
-class ProjectConfig {
-public:
-    std::string name;
-    std::string path;
-    CompilerConfig compiler;
-    DebuggerConfig debugger;
-    std::vector<std::string> libraries;
-};
